requirements.txt
Flask==2.3.2
Flask-HTTPAuth==4.8.0
Werkzeug==2.3.6
gunicorn==22.0.0
Js2Py==0.74
python-dotenv==1.0.0
websocket-client==1.7.0
websockets==10.4
undetected-chromedriver==3.5.5
webdriver-manager==4.0.1
paddleocr==2.7.0.3
paddle-bfloat==0.1.7
paddlepaddle==2.5.2
pandas==2.0.3 
Pillow==10.1.0
requests==2.31.0
tls-client==1.0.1
2captcha-python==1.2.0
anticaptchaofficial==1.0.52
bs4==0.0.1
pyreqwest_impersonate==0.4.8
undetected-chromedriver==3.5.5
undetected-playwright==0.2.0
scalene==1.5.31.1
scikit-image==0.21.0
scipy==1.10.1
scikit-learn==1.3.2
toolz==0.12.0
tqdm==4.66.1
nodriver==0.31